implement testcases and divided tests into separate files
