+++ mysite/views.py
@@ -10,10 +10,6 @@ def signup(request):
         form = UserCreationForm(request.POST)
         if form.is_valid():
             form.save()
-            username = form.cleaned_data.get('username')
-            raw_passwd = form.cleaned_data.get('password')
-            user = authenticate(username=username, password=raw_passwd)
-            login(request, user)
             messages.error(request, "Create account success! Please login!!!")
             return redirect("login")
     else:
