--- polls/tests.py
@@ -1,146 +1,0 @@
-import datetime
-from django.test import TestCase
-from django.utils import timezone
-from django.urls import reverse
-from .models import Question
-
-
-def create_question(question_text, days, end=None):
-    """Create a question with the given text and published day offset to now"""
-    time = timezone.now() + datetime.timedelta(days=days)
-    if end is not None:
-        end_time = timezone.now() + datetime.timedelta(days=end)
-        return Question.objects.create(question_text=question_text, pub_date=time, end_date=end_time)
-    return Question.objects.create(question_text=question_text, pub_date=time)
-
-
-class QuestionModelTests(TestCase):
-    def test_was_published_recently_with_future_question(self):
-        """Returns False for questions whose pub_date is in the future."""
-        time = timezone.now() + datetime.timedelta(days=30)
-        future_question = Question(pub_date=time)
-        self.assertIs(future_question.was_published_recently(), False)
-
-    def test_was_published_recently_with_old_question(self):
-        """Return False for question whose pub_date is older than 1 day."""
-        time = timezone.now() - datetime.timedelta(days=1, seconds=1)
-        old_question = Question(pub_date=time)
-        self.assertIs(old_question.was_published_recently(), False)
-
-    def test_was_published_recently_with_recent_question(self):
-        """Returns True for questions whose pub_date is within the last day."""
-        time = timezone.now() - datetime.timedelta(hours=23, minutes=59,
-                                                   seconds=59)
-        recent_question = Question(pub_date=time)
-        self.assertIs(recent_question.was_published_recently(), True)
-
-    def test_is_published_with_past_question(self):
-        """If question already published, `is_published` return true."""
-        question = create_question("", -4)
-        self.assertIs(question.is_published(), True)
-
-    def test_is_published_with_future_question(self):
-        """If question is unpublished, `is_published` return false."""
-        question = create_question("", 4)
-        self.assertIs(question.is_published(), False)
-
-    def test_can_vote_on_voting_period(self):
-        """If question is on voting period, `can_vote` return true."""
-        # poll with end date.
-        question = create_question("", 0, 10)
-        self.assertIs(question.can_vote(), True)
-
-        # poll with out end date.
-        question = create_question("", 0)
-        self.assertIs(question.can_vote(), True)
-
-    def test_can_not_vote_after_end_date(self):
-        """After question's end date, `can_vote` return false."""
-        question = create_question("", -10, -1)
-        self.assertIs(question.can_vote(), False)
-
-    def test_can_not_vote_on_unpublished_poll(self):
-        """If question is unpublished, `can_vote` return false."""
-        question = create_question("", 2)
-        self.assertIs(question.can_vote(), False)
-
-
-class QuestionIndexViewTests(TestCase):
-    def test_no_question(self):
-        """An message is displayed, if no question exist."""
-        response = self.client.get(reverse("polls:index"))
-        self.assertIs(response.status_code, 200)
-        self.assertContains(response, "No polls are available.")
-        self.assertQuerysetEqual(response.context["latest_question_list"], [])
-
-    def test_past_question(self):
-        """Questions with a pub_date in the past are displayed on the
-        index page.
-        """
-        question = create_question(question_text="Past question", days=-30)
-        response = self.client.get(reverse("polls:index"))
-        self.assertQuerysetEqual(response.context["latest_question_list"], [question], )
-
-    def test_future_question(self):
-        """Questions with a pub_date in the future are not displayed on the
-        index page.
-        """
-        create_question(question_text="Future question.", days=30)
-        response = self.client.get(reverse('polls:index'))
-        self.assertContains(response, "No polls are available.")
-        self.assertQuerysetEqual(response.context['latest_question_list'], [])
-
-    def test_future_question_and_past_question(self):
-        """Even if both past and future questions exist, only past questions
-        are displayed.
-        """
-        question = create_question(question_text="Past question.", days=-30)
-        create_question(question_text="Future question.", days=30)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(response.context['latest_question_list'], [question], )
-
-    def test_two_past_questions(self):
-        """The questions index page may display multiple questions."""
-        question1 = create_question(question_text="Past question 1.", days=-30)
-        question2 = create_question(question_text="Past question 2.", days=-5)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(response.context['latest_question_list'], [question2, question1], )
-
-
-class QuestionDetailViewTests(TestCase):
-    def test_future_question(self):
-        """The detail view of a question with a pub_date in the future returns
-        a 404 not found.
-        """
-        future_question = create_question(question_text='Future question.', days=5)
-        url = reverse('polls:detail', args=(future_question.id,))
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 302)
-
-    def test_past_question(self):
-        """The detail view of a question with a pub_date in the past displays
-        the question's text.
-        """
-        past_question = create_question(question_text='Past Question.', days=-5)
-        url = reverse('polls:detail', args=(past_question.id,))
-        response = self.client.get(url)
-        self.assertContains(response, past_question.question_text)
-
-
-class QuestionResultsViewTests(TestCase):
-    def test_votes_are_counted_correctly(self):
-        """Vote count should display correctly on web."""
-        question = create_question(question_text="New question", days=0)
-        question.choice_set.create(choice_text="first choice", votes=1)
-        choice = question.choice_set.get(pk=1)
-        url = reverse('polls:results', args=(question.id,))
-        response = self.client.get(url)
-        self.assertContains(response, f"{choice.choice_text}")
-        self.assertContains(response, '<th class="data">1</th>')
-
-    def test_future_question_result(self):
-        """If question is not published yet, It should not show result page"""
-        question = create_question("Future", 3)
-        url = reverse('polls:results', args=(question.id,))
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 302)
